+++ Lab4FCM/src/main/java/co/edu/udea/gr06_20171compumovil/lab4/pojos/Event.java
@@ -9,7 +9,7 @@ public class Event {
 
     private String name;
     private String description;
-    private int score;
+    private float score;
     private String picture;
     private String resposible;
     private String date;
@@ -34,10 +34,10 @@ public class Event {
     }
 
     public int getScore() {
-        return score;
+        return (int) score;
     }
 
-    public void setScore(int score) {
+    public void setScore(float score) {
         this.score = score;
     }
